Added rounding to time of failure. DIKERNEL-51

+++ DiKErnel/Src/DiKErnel.FunctionLibrary/Public/Revetment.h
@@ -87,7 +87,7 @@ namespace DiKErnel::FunctionLibrary
              * \return The time of failure.
              *         Unit = [s]
              */
-            static double TimeOfFailure(
+            static int TimeOfFailure(
                 double durationInTimeStepFailure,
                 double startTime);
     };
